Refactored test cases simplification
